docs: Simplify the documentation
Simplify and remove obsoleted parts from the documentation.

If we do a root cause analysis on documentation, we reach to a conclusion that documentation is only needed, because things are not obvious.
So the best solution to this problem is not writing documentation, instead eliminating the complexity and such parts which need explanation or use such technologies, which are being already documented.

Maintaining documentation can be a huge burden, I haven't ever seen such a system where documentation sooner or later didn't diverge from the code. And if that happens, it leads to just an even bigger confusion.

So in this case if a developer looks at the root project, probably he'll quickly spot the `Makefile`, the `.envrc` and the `pyproject.toml`.
And after they took a look at these files, they'll already know how to install, test, run, format the code/project. Without any documentation.